Refactored orders

+++ db/migrate/20231117234816_create_bill_of_material_lines.rb
@@ -3,7 +3,7 @@ class CreateBillOfMaterialLines < ActiveRecord::Migration[7.1]
     create_table :bill_of_material_lines do |t|
       t.references :component, foreign_key: { to_table: 'items' }
       t.references :group, foreign_key: { to_table: 'items' }
-      t.decimal :quantity
+      t.integer :quantity
       t.references :bill_of_material, null: false, foreign_key: true
       t.timestamps
     end
